Polymarket/main_polymarket.py
"""
Polymarket Betting Bot - Main Entry Point
"""
import asyncio
import sys
import os

# Add parent directory to path for imports
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from bookiebeats_api_monitor_polymarket import BookieBeatsAPIMonitorPolymarket
from polymarket_client import PolymarketClient

# TODO: Import dashboard and market matcher once they're created
# from dashboard_polymarket import app, socketio
# from market_matcher_polymarket import MarketMatcherPolymarket


async def main():
    """Main entry point"""
    print("=" * 60)
    print("Polymarket Betting Bot")
    print("=" * 60)
    print()
    
    # Initialize clients
    polymarket_client = PolymarketClient()
    await polymarket_client.init()
    
    # Initialize BookieBeats monitor
    from dotenv import load_dotenv
    load_dotenv()
    
    auth_token = os.getenv('BOOKIEBEATS_AUTH_TOKEN')
    cookies = os.getenv('BOOKIEBEATS_COOKIES')
    
    monitor = BookieBeatsAPIMonitorPolymarket(
        auth_token=auth_token,
        cookies=cookies
    )
    
    # TODO: Initialize dashboard and market matcher
    # market_matcher = MarketMatcherPolymarket(polymarket_client)
    
    print("✅ All components initialized")
    print()
    print("⚠️  NOTE: This is a skeleton implementation.")
    print("   You still need to:")
    print("   1. Implement market matching logic")
    print("   2. Implement order placement logic")
    print("   3. Create dashboard_polymarket.py")
    print("   4. Test with actual Polymarket API")
    print()
    
    # Start monitoring
    await monitor.start()
    
    try:
        await monitor.monitor_loop()
    except KeyboardInterrupt:
        print("\n🛑 Shutting down...")
    finally:
        await monitor.stop()
        await polymarket_client.close()


if __name__ == "__main__":
    asyncio.run(main())
