Polymarket/requirements.txt
aiohttp>=3.9.0
aiohttp-socks>=0.8.0
websockets>=12.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.0
thefuzz>=0.19.0
python-Levenshtein>=0.23.0
requests>=2.31.0
